Make timing tests compare to upper / lower bounds
When running tests in parallel, we get intermittent failures of
timing related tests due to the elapsed time being outstide of the
allowed range.

However, this could just be due to thread scheduling issues, as
an indeterminate amount of time could pass between the recorded trace
and the standard library call which was used as "ground truth".

Now, I bracket the trace records on both sides with standard library
timing calls, to provide expected upper and lower bounds, which should
be more forgiving of randomness from thread scheduling.

A follow up to changes in:

    Avoid hard-coded comparisons for timing tests
    4b11717435c68e995c68169468b26c2550b9aae9