Fix trace tests that failed intermittently on Apple Silicon
These tests opened several trace scopes and asserted that
the time spent in those scopes were nonzero. However, on
Apple Silicon we occasionally measured zero time spent
in those scopes, which caused the tests to fail. Because
these scopes contained no real work and the timers on
Apple Silicon have relatively low resolution (~42 ns/tick),
this is actually a legitimate outcome on that platform.

To avoid the test failure, we sleep for a microsecond in
between each trace scope so that the traces have some nonzero
time to measure. An alternate solution would be to change
the asserts to allow for zero times but it seemed good
for these tests to verify behavior in the nonzero case.

(Internal change: 2256221)